requirements.txt
streamlit
asyncio  
nest_asyncio
pandas
sqlmodel
sqlalchemy
python-dotenv